backend/requirements.txt
fastapi[all]
sqlalchemy
alembic
psycopg2
python-dotenv
pyjwt 
passlib 
bcrypt
python-jose[cryptography]